Revert "Merge remote-tracking branch 'origin/main' into release-v1.7.1"
This reverts commit bd74cce86a5e610477cc68bd555a636bcd735ca3, reversing
changes made to d2f77e88bd9f9ce018a71e78c120274336e80580.
