+++ meilisearch/src/option.rs
@@ -564,11 +564,11 @@ impl Opt {
                     }
                     if self.ssl_require_auth {
                         let verifier = AllowAnyAuthenticatedClient::new(client_auth_roots);
-                        config.with_client_cert_verifier(Arc::from(verifier))
+                        config.with_client_cert_verifier(verifier)
                     } else {
                         let verifier =
                             AllowAnyAnonymousOrAuthenticatedClient::new(client_auth_roots);
-                        config.with_client_cert_verifier(Arc::from(verifier))
+                        config.with_client_cert_verifier(verifier)
                     }
                 }
                 None => config.with_no_client_auth(),
