+++ milli/src/vector/error.rs
@@ -2,7 +2,6 @@ use std::path::PathBuf;
 
 use hf_hub::api::sync::ApiError;
 
-use super::ollama::OllamaError;
 use crate::error::FaultSource;
 use crate::vector::openai::OpenAiError;
 
@@ -72,17 +71,6 @@ pub enum EmbedErrorKind {
     OpenAiRuntimeInit(std::io::Error),
     #[error("initializing web client for sending embedding requests failed: {0}")]
     InitWebClient(reqwest::Error),
-    // Dedicated Ollama error kinds, might have to merge them into one cohesive error type for all backends.
-    #[error("unexpected response from Ollama: {0}")]
-    OllamaUnexpected(reqwest::Error),
-    #[error("sent too many requests to Ollama: {0}")]
-    OllamaTooManyRequests(OllamaError),
-    #[error("received internal error from Ollama: {0}")]
-    OllamaInternalServerError(OllamaError),
-    #[error("model not found. Meilisearch will not automatically download models from the Ollama library, please pull the model manually: {0}")]
-    OllamaModelNotFoundError(OllamaError),
-    #[error("received unhandled HTTP status code {0} from Ollama")]
-    OllamaUnhandledStatusCode(u16),
 }
 
 impl EmbedError {
@@ -140,26 +128,6 @@ impl EmbedError {
 
     pub fn openai_initialize_web_client(inner: reqwest::Error) -> Self {
         Self { kind: EmbedErrorKind::InitWebClient(inner), fault: FaultSource::Runtime }
-    }
-
-    pub(crate) fn ollama_unexpected(inner: reqwest::Error) -> EmbedError {
-        Self { kind: EmbedErrorKind::OllamaUnexpected(inner), fault: FaultSource::Bug }
-    }
-
-    pub(crate) fn ollama_model_not_found(inner: OllamaError) -> EmbedError {
-        Self { kind: EmbedErrorKind::OllamaModelNotFoundError(inner), fault: FaultSource::User }
-    }
-
-    pub(crate) fn ollama_too_many_requests(inner: OllamaError) -> EmbedError {
-        Self { kind: EmbedErrorKind::OllamaTooManyRequests(inner), fault: FaultSource::Runtime }
-    }
-
-    pub(crate) fn ollama_internal_server_error(inner: OllamaError) -> EmbedError {
-        Self { kind: EmbedErrorKind::OllamaInternalServerError(inner), fault: FaultSource::Runtime }
-    }
-
-    pub(crate) fn ollama_unhandled_status_code(code: u16) -> EmbedError {
-        Self { kind: EmbedErrorKind::OllamaUnhandledStatusCode(code), fault: FaultSource::Bug }
     }
 }
 
@@ -224,13 +192,6 @@ impl NewEmbedderError {
         Self {
             kind: NewEmbedderErrorKind::CouldNotDetermineDimension(inner),
             fault: FaultSource::Runtime,
-        }
-    }
-
-    pub fn ollama_could_not_determine_dimension(inner: EmbedError) -> NewEmbedderError {
-        Self {
-            kind: NewEmbedderErrorKind::CouldNotDetermineDimension(inner),
-            fault: FaultSource::User,
         }
     }
