--- workloads/settings-proximity-precision.json
@@ -1,86 +1,0 @@
-{
-  "name": "settings-proximity-precision.json",
-  "run_count": 2,
-  "extra_cli_args": [
-    "--max-indexing-threads=4"
-  ],
-  "assets": {
-    "150k-people.json": {
-      "local_location": null,
-      "remote_location": "https://milli-benchmarks.fra1.digitaloceanspaces.com/bench/datasets/150k-people.json",
-      "sha256": "28c359a0956958af0ba204ec11bad3045a0864a10b4838914fea25a01724f84b"
-    }
-  },
-  "commands": [
-    {
-      "route": "indexes/peoples/settings",
-      "method": "PATCH",
-      "body": {
-        "inline": {
-          "searchableAttributes": [
-            "last_name",
-            "first_name",
-            "featured_job_organization_name",
-            "facebook_url",
-            "twitter_url",
-            "linkedin_url"
-          ],
-          "filterableAttributes": [
-            "city",
-            "region",
-            "country_code",
-            "featured_job_title",
-            "featured_job_organization_name"
-          ],
-          "dictionary": [
-            "https://",
-            "http://",
-            "www.",
-            "crunchbase.com",
-            "facebook.com",
-            "twitter.com",
-            "linkedin.com"
-          ],
-          "stopWords": [
-            "https://",
-            "http://",
-            "www.",
-            "crunchbase.com",
-            "facebook.com",
-            "twitter.com",
-            "linkedin.com"
-          ]
-        }
-      },
-      "synchronous": "DontWait"
-    },
-    {
-      "route": "indexes/peoples/documents",
-      "method": "POST",
-      "body": {
-        "asset": "150k-people.json"
-      },
-      "synchronous": "WaitForTask"
-    },
-    {
-      "route": "indexes/peoples/settings",
-      "method": "PATCH",
-      "body": {
-        "inline": {
-          "proximityPrecision": "byAttribute"
-        }
-      },
-      "synchronous": "WaitForTask"
-    },
-    {
-      "route": "indexes/peoples/settings",
-      "method": "PATCH",
-      "body": {
-        "inline": {
-          "proximityPrecision": "byWord"
-        }
-      },
-      "synchronous": "WaitForTask"
-    }
-  ]
-}
